Problem described by Rafal Lalik:
When I use fit tool and peak finder, after use I have list of polX and gaus functions. When I click on gaus then I see in histogram position of fitted peak related to this function. But when I use arrows to navigate in this list, then peak position is not displayed/refreshed. I don't know how it is done, but probably it will be needed to change QList object event from ObMouseClick to OnSelect or smth like this.

I think it would be good if this feature (browsing with arrows) was implemented.

Solution:
Connect slot to other signal, which indicates change of current item in QListWidget

git-svn-id: https://subversion.gsi.de/go4/trunk@644 302e7798-3865-0410-85c9-d38765bbeff6